Refactor scripts
Reorganized scripts to reduce total number of files, avoid code
repetition, and make script purposes clear.  Should be no change in
function

Several scripts were combined and for now, the original comments are
left and file names are also left in comments.

This fixes #11 and works towards the improvements suggested in #12